Move the creation of command aliases outside of init() now that the building of comands doesn't create the http clients.
Moved EnsureConfiguration() into config.go

Removed print lines
